requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.34.127
botocore==1.34.127
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.18
django-admin-honeypot==1.1.0
django-admin-ordering==0.14.2
django-admin-sortable2==2.2.1
django-baton==3.1.0
django-ckeditor==6.7.1
django-colorful==1.3
django-cors-headers==3.8.0
django-formtools==2.5.1
django-health-check==3.18.2
django-import-export==4.0.8
django-js-asset==2.2.0
django-otp==1.5.0
django-password-strength==1.2.1
django-password-validators==1.7.3
django-phonenumber-field==7.3.0
django-reversion==5.0.12
django-storages==1.14.3
django-subject-imagefield==0.2.13
django-two-factor-auth==1.16.0
django-zxcvbn-password==2.1.1
django_csp==3.8
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openpyxl==3.1.4
packaging==24.1
phonenumberslite==8.13.39
pillow==10.3.0
protobuf==4.21.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-resize-image==1.1.20
pytz==2024.1
PyYAML==6.0
qrcode==7.4.2
requests==2.32.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
xlrd==2.0.1
xlwt==1.3.0
zxcvbn==4.4.28
